requirements.txt
Flask==3.0.3
google-generativeai==0.6.0
mysql-connector-python==8.4.0
pillow==10.3.0